feat: add decorator example
