+++ packages/caption/caption.stories.js
@@ -1,4 +1,4 @@
-import showcase from "./caption.showcase";
 import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./caption.showcase";
 
 showcaseConverter(module, showcase);
